Update styling for check in and checkout pages
Most of the work is on the checkin page but includes some margin for the
checkout page.

This switches checkin over to a CSS grid based layout so that
everything lines up nicely.

Also includes the ability to run npm watch to autobuild changes

Includes some settings changes so that review apps work without
modifying the ALLOWED_HOSTS